Rework base class methods into service methods
