+++ tools/Azure.Mcp.Tools.SreAgent/src/Commands/Threads/ThreadsCreateCommand.cs
@@ -13,7 +13,7 @@ using Microsoft.Mcp.Core.Models.Command;
 namespace Azure.Mcp.Tools.SreAgent.Commands.Threads;
 
 [CommandMetadata(Id = "c81d2450-f262-40a4-a9b0-5b2e1c3d0e03", Name = "create", Title = "Create Thread", Description = "Create a new thread on an SRE Agent and start a conversation by sending the opening message. Returns the initial agent response.", Destructive = false, Idempotent = false, OpenWorld = true, ReadOnly = false, Secret = false, LocalRequired = false)]
-public sealed class ThreadsCreateCommand(ILogger<ThreadsCreateCommand> logger, ISreAgentService sreAgentService) : ThreadsCommandBase<ThreadsCreateOptions>
+public sealed class ThreadsCreateCommand(ILogger<ThreadsCreateCommand> logger, ISreAgentService sreAgentService) : SreAgentDataPlaneCommand<ThreadsCreateOptions>
 {
     private readonly ILogger<ThreadsCreateCommand> _logger = logger;
     private readonly ISreAgentService _sreAgentService = sreAgentService;
@@ -41,10 +41,10 @@ public sealed class ThreadsCreateCommand(ILogger<ThreadsCreateCommand> logger, I
         try
         {
             var endpoint = await ResolveEndpointAsync(_sreAgentService, options, cancellationToken);
-            var thread = await _sreAgentService.CreateThreadAsync(endpoint, CreateThreadRequest(options.Message!, options.Agent!), options.Tenant, cancellationToken);
+            var thread = await _sreAgentService.CreateThreadAsync(endpoint, SreAgentCommandHelpers.CreateThreadRequest(options.Message!, options.Agent!), options.Tenant, cancellationToken);
             var threadId = thread?.Id;
-            var messages = string.IsNullOrWhiteSpace(threadId) ? [] : await PollForCompletionAsync(_sreAgentService, endpoint, threadId, options.Tenant, TimeSpan.FromMinutes(2), false, cancellationToken);
-            context.Response.Results = ResponseResult.Create(new SreAgentThreadOperationResult(threadId, "created", messages), SreAgentJsonContext.Default.SreAgentThreadOperationResult);
+            var messages = string.IsNullOrWhiteSpace(threadId) ? [] : await _sreAgentService.PollThreadForCompletionAsync(endpoint, threadId, options.Tenant, TimeSpan.FromMinutes(2), false, cancellationToken);
+            context.Response.Results = ResponseResult.Create(new(threadId, "created", messages), SreAgentJsonContext.Default.SreAgentThreadOperationResult);
         }
         catch (Exception ex)
         {
