+++ tools/Azure.Mcp.Tools.SreAgent/src/Commands/Threads/ThreadsInvestigateCommand.cs
@@ -12,7 +12,7 @@ using Microsoft.Mcp.Core.Models.Command;
 namespace Azure.Mcp.Tools.SreAgent.Commands.Threads;
 
 [CommandMetadata(Id = "ab73d6fa-d53e-446c-9d4c-9d8cf41a3106", Name = "investigate", Title = "Investigate With Agent", Description = "Investigate an issue or incident using an SRE Agent. Sends your investigation message and automatically follows up on agent questions until the investigation is complete.", Destructive = false, Idempotent = false, OpenWorld = true, ReadOnly = false, Secret = false, LocalRequired = false)]
-public sealed class ThreadsInvestigateCommand(ILogger<ThreadsInvestigateCommand> logger, ISreAgentService sreAgentService) : ThreadsCommandBase<ThreadsInvestigateOptions>
+public sealed class ThreadsInvestigateCommand(ILogger<ThreadsInvestigateCommand> logger, ISreAgentService sreAgentService) : SreAgentDataPlaneCommand<ThreadsInvestigateOptions>
 {
     private readonly ILogger<ThreadsInvestigateCommand> _logger = logger;
     private readonly ISreAgentService _sreAgentService = sreAgentService;
@@ -43,7 +43,7 @@ public sealed class ThreadsInvestigateCommand(ILogger<ThreadsInvestigateCommand>
         var options = BindOptions(parseResult);
         try
         {
-            var result = await RunInvestigationAsync(_sreAgentService, options, autoApprove: false, cancellationToken);
+            var result = await _sreAgentService.RunInvestigationAsync(options, autoApprove: false, cancellationToken);
             context.Response.Results = ResponseResult.Create(result, SreAgentJsonContext.Default.SreAgentInvestigationResult);
         }
         catch (Exception ex)
