+++ tools/Azure.Mcp.Tools.SreAgent/src/Commands/Threads/ThreadsListCommand.cs
@@ -11,7 +11,7 @@ using Microsoft.Mcp.Core.Models.Command;
 namespace Azure.Mcp.Tools.SreAgent.Commands.Threads;
 
 [CommandMetadata(Id = "23b7c0d6-29c5-4d8f-82de-1e0edc1a9b01", Name = "list", Title = "List Threads", Description = "List SRE Agent chat threads.", Destructive = false, Idempotent = true, OpenWorld = false, ReadOnly = true, Secret = false, LocalRequired = false)]
-public sealed class ThreadsListCommand(ILogger<ThreadsListCommand> logger, ISreAgentService sreAgentService) : ThreadsCommandBase<ThreadsListOptions>
+public sealed class ThreadsListCommand(ILogger<ThreadsListCommand> logger, ISreAgentService sreAgentService) : SreAgentDataPlaneCommand<ThreadsListOptions>
 {
     private readonly ILogger<ThreadsListCommand> _logger = logger;
     private readonly ISreAgentService _sreAgentService = sreAgentService;
